+++ both/components/CollapseBox.jsx
@@ -20,9 +20,10 @@ CollapseBox = React.createClass({
 			    <h4 onClick={this.toggle} className="background" style={{margin:"10px 15px",cursor:"pointer"}}>
 			    	<span><i className={"fa fa-caret-"+(collapsed?'right':'down')}></i> {this.props.title}</span>
 			    </h4>
+			    {collapsed?null:
 			    <div className={"collapse-box "+(collapsed?'collapsed':'')} ref="collapser">
 			    	{this.props.children}
-			    </div>
+			    </div>}
 			</div>
 		)
 	}